TINY-10324: remove code duplication
